Split up logging module
